% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dev}
\alias{dev}
\alias{dev<-}
\title{Mark operator as deviation measure}
\usage{
dev(op)

dev (op) <- value
}
\arguments{
\item{op}{the operator (function)}

\item{value}{logical indicating the operator type}
}
\value{
logical indicating the type of operator. \code{NULL} if the attribute is missing.
}
\description{
The operators measure either a performance (i.e. accordance between reference and prediction) or
a deviation. \code{dev (op) == TRUE} marks operators measuring deviation.
}
\examples{

dev (wRMSE)
myop <- function (r, p) p * (r == 1)
dev (myop) <- TRUE

}
\seealso{
\code{\link{sens}} \code{\link{postproc}}
}
\author{
Claudia Beleites
}
