% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2matrix.R
\name{factor2matrix}
\alias{factor2matrix}
\title{Convert hard class labels to membership matrix}
\usage{
factor2matrix(f)
}
\arguments{
\item{f}{factor with class labels}
}
\value{
matrix of size \code{length (f)} x \code{nlevels (f)}
}
\description{
Converts a factor with hard class memberships into a membership matrix
}
\seealso{
\code{\link[softclassval]{hardclasses}} for the inverse
}
\author{
Claudia Beleites
}
