% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardclasses.R
\name{hard}
\alias{hard}
\alias{hard<-}
\title{Mark operator as hard measure}
\usage{
hard(op)

hard (op) <- value
}
\arguments{
\item{op}{the operator (function)}

\item{value}{logical indicating the operator type}
}
\value{
logical indicating the type of operator. \code{NULL} if the attribute is missing.
}
\description{
The operators may work only for hard classes (see \code{\link[softclassval:operators]{and}}). \code{hard (op)
== TRUE} marks hard operators.
}
\examples{

hard (and)
myop <- function (r, p) p * (r == 1)
hard (myop) <- TRUE

}
\seealso{
\code{\link{sens}} \code{\link[softclassval:operators]{and}}
}
\author{
Claudia Beleites
}
