% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softclassval-package.R
\docType{package}
\name{softclassval-package}
\alias{softclassval}
\alias{softclassval-package}
\title{Soft classification performance measures}
\description{
Extension of sensitivity, specificity, positive and negative predictive value to continuous
predicted and reference memberships in [0, 1].
}
\seealso{
Useful links:
\itemize{
  \item \url{http://softclassval.r-forge.r-project.org/}
}

}
\author{
Claudia Beleites
}
\keyword{internal}
