% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unittests.R
\name{softclassval.unittest}
\alias{softclassval.unittest}
\title{Run the unit tests}
\usage{
softclassval.unittest()
}
\value{
invisibly \code{TRUE} if the tests pass, \code{NA} if \link[svUnit]{svUnit} is not
available. Stops if errors are encountered.
}
\description{
Run the unit tests attached to the functions via \link[svUnit]{svUnit}
}
\seealso{
\link[svUnit]{svUnit}
}
\author{
Claudia Beleites
}
\keyword{programming}
\keyword{utilities}
