\name{softclassval-package}
\alias{softclassval-package}
\alias{softclassval-package}
\title{Soft classification performance measures...}
\description{Soft classification performance measures}
\details{Extension of sensitivity, specificity, positive and negative predictive value to continuous
predicted and reference memberships in [0, 1]}
\docType{package}

