% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softclassval.R
\docType{package}
\name{softclassval-package}
\alias{softclassval-package}
\title{Soft classification performance measures}
\description{
Extension of sensitivity, specificity, positive and negative predictive value to continuous
predicted and reference memberships in [0, 1].
}
\author{
C. Beleites
}

