\name{softclassval.unittest}
\alias{softclassval.unittest}
\title{Run the unit tests}
\usage{
  softclassval.unittest()
}
\value{
  invisibly \code{TRUE} if the tests pass, \code{NA} if
  \link[svUnit]{svUnit} is not available. Stops if errors
  are encountered.
}
\description{
  Run the unit tests attached to the functions via
  \link[svUnit]{svUnit}
}
\author{
  Claudia Beleites
}
\seealso{
  \link[svUnit]{svUnit}
}

